Push measurements serially
Simplify the code significantly by avoid curl multi handles. There was a
bug around headers and instead of attempting a fix, just switch to a
simple serial push of batches. Since the number of metrics generated by
apps using this library is expected to be relatively low, this helps
keep the code simple without loss of functionality.