Use the c-ares library to avoid problems with curl (#26)
By default libcurl implements timeouts on top of `SIGALRM` for name
resolution. This can cause segment violations due to the use of unsafe
operations inside the signal handler (in particular `longjmp`).  With
this PR we ship a libcurl (7.64.1) compiled with c-ares enabled to make
name resolutions async and be able to deal with timeouts in DNS queries
without resorting to signals.