Avoid copies when POSTing JSON
Add an overload to `HttpClient::Post` that takes a shared pointer to a
buffer allocated by the user. This allows us to get a string from a JSON
doc using a custom allocator that saves it to memory managed by a shared
pointer, and therefore we can avoid the mem copy for this common use
case.