Option for reading the response body for http reqs (#40)
Some uses of the http client require the body of the response from the
server for diagnostic purposes. This changes the signature of the `Post`
methods of the HTTP client to return an `HttpResponse` type instead of
the simple status code. The response includes the raw body as a string
(i.e. no decoding based on the encoding header from the server) and
a map of headers. Both are optional in case the client is not
interested.