Benchmark different hash tables for the Registry
Try different hash tables as possible replacements for the
`ska::flat_hash_map` used for the Registry (it's also used for Tags but
the usage of that hash table is different and will need a different
benchmark)

The tests show a significant performance benefit to switching to
`tsl::hopscotch_map` while enabling storing the hash code in the table.

While adding these dependencies I removed `pcre` which is not used by
spectator_cpp.