Keep track of invalid measurements (#67)
Show the unique warnings we get from the server when we post
measurements. Additionally standardize on the Ids used by spectator-js:

`spectator.measurements id=sent|dropped
error=validation|http-error|other`

for the counters that track the number of measurements sent successfully
versus the ones that were dropped for various reasons.