Add support for eager percentile metrics
Allow percentile timers and distribution summaries to register all
counters that are used for buckets at creation time in addition to the
lazy option where we get and register counters when Record is called.

This should allow applications to opt for a more expensive creation
operation in exchange for making Record much faster.