Add support for eager percentile metrics (#70)
* Add support for eager percentile metrics

Allow percentile timers and distribution summaries to register all
counters that are used for buckets at creation time in addition to the
lazy option where we get and register counters when Record is called.

This should allow applications to opt for a more expensive creation
operation in exchange for making Record much faster.

* Benchmark to test policies for percentiles

The benchmark shows that the eager policy is much more expensive than
the lazy one we were using. So remove the eager_policy to avoid users
mistakingly using it.

The code is cleaner with this policy separation
since it allows us to reduce code duplication from percentile
timers/dist summaries but there's currently no benefit from moving away
from the lazy policy of creating the counters once and caching their
value.