Adds stateless/stateful registries (#78)
A stateless registry is usually used in conjunction with spectatord and
does not require to keep state on the current process. Tests done by
users of this library are greatly simplified if the registry and meters
can be queried though, therefore requiring some local state.

This change adds a way to compose whether the registry will be
stateless/stateful and which publisher will be used to send all updates
in the stateless case.