fix number formatting (#94)
When setting AgeGauges to epoch seconds, such as `1671644992`, the
resulting protocol line was `A:gauge:1.67164e+09`, when it is expected
to be `A:gauge:1671644992`. The original AgeGauge tests were not setting
large enough numbers to discover this formatting issue.

Swapping `StrCat` for `StrFormat` and stripping trailing zeroes and decimal
places from the resulting number strings produces the output we want. The
AgeGauge tests have been updated to use more representative numbers.

This change also locks Bazel at version 5.4.0, since there are a number of
breaking changes in the 6.0.0 release.