switch build to conan 1.x (#100)
The Conan 1.x package manager for C++ is a bit easier to work with than the
Bazel system. This allows us to more easily update dependencies, to keep them
in-sync with the atlas-system-agent project, which also uses Conan.

This change includes some additional configuration in a profile, which is used
to inject necessary cxxflags to the package builds, so that the local debug
builds can be run with the address sanitizer enabled. Without this additional
setting, segfaults will occur during testing, because the address sanitizer
will not be applied to both the project and the dependencies.

Switching back and forth between Debug and Release builds requires removing
all packages and builds from the Conan cache, due to the way these flags are
injected. The `./build.sh clean` script shows how this is done.